Create HomePage
Create HomePage class to represent the homepage as an object. Add method to go to login page which returns a LoginPage object. Add method to get the current url and return it as a string. Add method to logout by clicking the logout button in the user menu. Add a method to click the login button from the homepage.