Create LoginPage.cs
Create LoginPage to be an object representation of the login page. Add method to login by passing in a string email and string password. Add method to mark the Remember Me checkbox. Add method to get login information that has been remembered and return it as a tuple of two strings. Add method to return an error message when login fails.